fix: change SensorStatus.batteryPercent from Int? to Double?
The Medtrum API returns batteryPercent as a Double (e.g. 0.91)
which caused Kotlin serialization to crash with 'Unexpected symbol
in numeric literal', making the entire status response unparseable
and showing 'No sensor data'.

+++ shared/src/main/java/com/nimbleflux/glucosesync/shared/api/model/ApiModels.kt
@@ -42,7 +42,7 @@ data class SensorStatus(
     val serial: Long? = null,
     val status: Int? = null,
     val sensorLifetimeTotalCount: Int? = null,
-    val batteryPercent: Int? = null
+    val batteryPercent: Double? = null,
 )
 
 @Serializable
